+++ compiler/include/instruction_set.hpp
@@ -1,29 +1,34 @@
 #ifndef INSTRUCTION_SET_HPP
 #define INSTRUCTION_SET_HPP
 
+#include <cmath>
 #include <fstream>
 #include <iostream>
 #include <map>
 #include <memory>
 #include <string>
 #include <vector>
-#include <cmath>
 
 #include "ast.hpp"
-#include "symbol_table.hpp"
 #include "register_allocator.hpp"
+#include "symbol_table.hpp"
 #include "utils.hpp"
 
 class InstructionSet {
     std::ofstream& out;
     int current_stack_depth = 0;
-public:
+
+   public:
     InstructionSet(std::ofstream& o) : out(o) {}
 
     // Instruction set
     std::string get_size_prefix(int size);
-    int get_stack_depth() const { return current_stack_depth; }
-    void reset_stack_depth() { current_stack_depth = 0; }
+    int get_stack_depth() const {
+        return current_stack_depth;
+    }
+    void reset_stack_depth() {
+        current_stack_depth = 0;
+    }
     bool isAFloatingPoint(const TypeNode* type);
 
     void emit(const std::string& instr);
@@ -45,7 +50,8 @@ public:
     void emit_lea_stack(int offset);
     void emit_load_constant(const std::string& instr, const std::string& label);
     void emit_dereference();
-    void emit_data_entry(const std::string& label, const std::string& type, const std::string& value);
+    void emit_data_entry(const std::string& label, const std::string& type,
+                         const std::string& value);
 
     void emit_adv(int size, const TypeNode* type, const std::string& base_reg, int offset,
                   const std::string& src_val);
@@ -54,6 +60,7 @@ public:
     void load_from_address(int size, const std::string& reg);
     void emit_print(int size, const std::shared_ptr<TypeNode>& type);
     void emit_print_int(const std::string& reg);
+    void emit_print_raw(int size, const std::shared_ptr<TypeNode>& type);
 
     void emit_binary_op(const std::string& op_instr, char type);
 };