+++ compiler/src/register_allocator.cpp
@@ -1,31 +1,37 @@
 #include "register_allocator.hpp"
-#include "utils.hpp"
 
 #include <algorithm>
 
-RegisterAllocator::RegisterAllocator() {
-    register_pool = { R10, R11, RBX, RCX, RDI, RSI };
+#include "utils.hpp"
 
-    register_lookup[R10] = { "r10b", "r10d", "r10" };
-    register_lookup[R11] = { "r11b", "r11d", "r11" };
-    register_lookup[RBX] = { "bl",   "ebx",  "rbx" };
-    register_lookup[RCX] = { "cl",   "ecx",  "rcx" };
-    register_lookup[RDI] = { "di",   "edi",  "rdi" };
-    register_lookup[RSI] = { "si",   "esi",  "rsi" };
+RegisterAllocator::RegisterAllocator() {
+    register_pool = {R10, R11, RBX, RCX, RDI, RSI};
+
+    register_lookup[R10] = {"r10b", "r10d", "r10"};
+    register_lookup[R11] = {"r11b", "r11d", "r11"};
+    register_lookup[RBX] = {"bl", "ebx", "rbx"};
+    register_lookup[RCX] = {"cl", "ecx", "rcx"};
+    register_lookup[RDI] = {"di", "edi", "rdi"};
+    register_lookup[RSI] = {"si", "esi", "rsi"};
 }
 
 std::string RegisterAllocator::get_name(RegID reg, int byte_size) {
     const RegStrings& bundle = register_lookup[reg];
 
     switch (byte_size) {
-        case 1: return bundle.byte1;
-        case 4: return bundle.byte4;
-        default: return bundle.byte8;
+        case 1:
+            return bundle.byte1;
+        case 4:
+            return bundle.byte4;
+        default:
+            return bundle.byte8;
     }
 }
 
 RegisterAllocator::RegID RegisterAllocator::allocate() {
-    if (register_pool.empty()) { Utils::report_error("Register Allocator", "Register pool is empty."); }
+    if (register_pool.empty()) {
+        Utils::report_error("Register Allocator", "Register pool is empty.");
+    }
 
     RegID allocated_register = register_pool.back();
     register_pool.pop_back();
@@ -42,6 +48,6 @@ void RegisterAllocator::free_reg(RegID reg) {
 }
 
 void RegisterAllocator::reset() {
-    register_pool = { R10, R11, RBX, RCX, RDI, RSI };
+    register_pool = {R10, R11, RBX, RCX, RDI, RSI};
     return;
 }